+++ public/js/mambo-ui-0.0.4-stories.js
@@ -1993,14 +1993,21 @@ function storyTimePicker(selectedStory) {
   ui.timePicker(config);
 }
 function storyToaster(selectedStory) {
-  const activeToasters = {};
+  const activeToasters = {
+    closeIfExists: function(toasterId) {
+      if (this[toasterId]) {
+        this[toasterId].close();
+        delete this[toasterId];
+        return true;
+      }
+      return false;
+    }
+  };
   const configButton = {
     text: "Tooltip with Mambo Button",
     parentTag: selectedStory.parentTag,
     onClick: () => {
-      if (activeToasters["toaster-with-mambo-button"]) {
-        activeToasters["toaster-with-mambo-button"].close();
-        delete activeToasters["toaster-with-mambo-button"];
+      if (activeToasters.closeIfExists("toaster-with-mambo-button")) {
         return;
       }
       let toasterConfig = {
@@ -2008,12 +2015,12 @@ function storyToaster(selectedStory) {
         closeButton: false,
         open: true,
         message: "Mambo Tooltip Successfully Completed",
-        autoHideDuration: 5e3,
         size: "large",
         type: "success",
-        onClose: (context) => {
-          context.Toaster.close();
-          delete activeToasters["toaster-with-mambo-button"];
+        onClosed: (context) => {
+          if (activeToasters["toaster-with-mambo-button"]) {
+            delete activeToasters["toaster-with-mambo-button"];
+          }
         },
         onComplete: (context) => {
           context.Toaster.getBodyTag().innerHTML = `
@@ -2023,18 +2030,25 @@ function storyToaster(selectedStory) {
 									</div>
 									<mambo-button style="padding-left: 1rem; pointer-events: auto;" id='btn-undo-mambo-button'></mambo-button>`;
           const buttonUndo = document.getElementById("btn-undo-mambo-button");
-          const timeOut = setTimeout(() => {
-            context.Toaster.close();
-            delete activeToasters["toaster-with-mambo-button"];
-          }, context.Toaster.autoHideDuration());
+          let timeout2 = null;
+          if (!toasterConfig.autoHideDuration) {
+            timeout2 = setTimeout(() => {
+              context.Toaster.close();
+            }, 1500);
+          }
           buttonUndo.setup({
             text: "Cancel",
             type: "secondary",
             size: "small",
             onClick: () => {
-              clearTimeout(timeOut);
+              if (!toasterConfig.autoHideDuration) {
+                clearTimeout(timeout2);
+                timeout2 = null;
+              }
               context.Toaster.close();
-              delete activeToasters["toaster-with-mambo-button"];
+              if (activeToasters["toaster-with-mambo-button"]) {
+                delete activeToasters["toaster-with-mambo-button"];
+              }
             }
           });
         }
@@ -2043,26 +2057,32 @@ function storyToaster(selectedStory) {
     }
   };
   ui.button(configButton);
-  const normalVariants = [{ h: "center", v: "top", type: "success", size: "medium", animation: "top-bottom" }];
+  const normalVariants = [
+    {
+      h: "center",
+      v: "top",
+      type: "success",
+      size: "medium",
+      animation: "top-bottom",
+      autoHideDuration: 1500
+    }
+  ];
   const persistentVariants = [
-    { h: "left", v: "top", type: "info", size: "small", animation: "top-bottom" },
-    { h: "center", v: "top", type: "success", size: "medium", animation: "top-bottom" },
-    { h: "right", v: "top", type: "error", size: "large", animation: "top-bottom" },
-    { h: "left", v: "center", type: "warning", size: "small", animation: "top-bottom" },
-    { h: "center", v: "center", type: "info", size: "medium", animation: "top-bottom" },
-    { h: "right", v: "center", type: "success", size: "large", animation: "top-bottom" },
-    { h: "left", v: "bottom", type: "error", size: "small", animation: "bottom-top" },
-    { h: "center", v: "bottom", type: "warning", size: "medium", animation: "bottom-top" },
-    { h: "right", v: "bottom", type: "success", size: "large", animation: "bottom-top" }
+    { h: "left", v: "top", type: "info", size: "small", animation: "top-bottom", autoHideDuration: 1500 },
+    { h: "right", v: "top", type: "error", size: "large", animation: "top-bottom", autoHideDuration: 1500 },
+    { h: "left", v: "center", type: "warning", size: "small", animation: "top-bottom", autoHideDuration: 1500 },
+    { h: "center", v: "center", type: "info", size: "medium", animation: "top-bottom", autoHideDuration: 1500 },
+    { h: "right", v: "center", type: "success", size: "large", animation: "top-bottom", autoHideDuration: 1500 },
+    { h: "left", v: "bottom", type: "error", size: "small", animation: "bottom-top", autoHideDuration: 1500 },
+    { h: "center", v: "bottom", type: "warning", size: "medium", animation: "bottom-top", autoHideDuration: 1500 },
+    { h: "right", v: "bottom", type: "success", size: "large", animation: "bottom-top", autoHideDuration: 1500 }
   ];
   normalVariants.forEach((variant) => {
     const config = {
       text: `Toaster ${variant.h}-${variant.v}`,
       parentTag: selectedStory.parentTag,
       onClick: () => {
-        if (activeToasters[`${variant.h}-${variant.v}`]) {
-          activeToasters[`${variant.h}-${variant.v}`].close();
-          delete activeToasters[`${variant.h}-${variant.v}`];
+        if (activeToasters.closeIfExists(`${variant.h}-${variant.v}`)) {
           return;
         }
         let toasterConfig = {
@@ -2074,20 +2094,21 @@ function storyToaster(selectedStory) {
 	 							 <span style="font-size: 1rem; font-weight: 600;">Title:</span>
 	 							 <span style="font-size: 0.85rem; font-weight: 400;">toaster: ${variant.h}-${variant.v}</span>
 	 						    </div>`,
-          autoHideDuration: 1e3,
+          autoHideDuration: variant.autoHideDuration,
           type: variant.type,
           size: variant.size,
           animation: { name: variant.animation },
           persist: false,
           onClose: (context) => {
             context.Toaster.close();
+            console.log("Toaster Button Close");
             delete activeToasters[`${variant.h}-${variant.v}`];
           },
           onComplete: (context) => {
-            const timeOut = setTimeout(() => {
-              context.Toaster.close();
-              delete activeToasters[`${variant.h}-${variant.v}`];
-            }, context.Toaster.autoHideDuration());
+          },
+          onClosed: (context) => {
+            console.log("Toaster Closed");
+            delete activeToasters[`${variant.h}-${variant.v}`];
           }
         };
         activeToasters[`${variant.h}-${variant.v}`] = ui.toaster(toasterConfig);
@@ -2096,29 +2117,25 @@ function storyToaster(selectedStory) {
     ui.button(config);
   });
   let toasterInstance = null;
-  let currentTimeout = null;
+  let timeout = null;
   const createToaster = (context) => {
     const baseConfig = {
       message: `<div style="display: flex; flex-direction: column;">
 								<span style="font-size: 0.85rem; font-weight: 400;">Persistent Toaster Created!</span>
 								</div>`,
       closeButton: true,
-      autoHideDuration: 1e3,
+      autoHideDuration: 500,
       open: context.open,
       persist: true,
       distance: { y: "1rem", x: "0rem" },
       onClose: (context2) => {
+        console.log("Toaster Button Close");
         context2.Toaster.close();
-        if (currentTimeout) {
-          clearTimeout(currentTimeout);
-          currentTimeout = null;
-        }
       },
       onComplete: (context2) => {
-        currentTimeout = setTimeout(() => {
-          context2.Toaster.close();
-          currentTimeout = null;
-        }, context2.Toaster.autoHideDuration());
+      },
+      onClosed: (context2) => {
+        console.log("Toaster Closed");
       }
     };
     toasterInstance = ui.toaster(baseConfig);
@@ -2129,13 +2146,16 @@ function storyToaster(selectedStory) {
       text: `Persistent Toaster ${variant.h}-${variant.v}`,
       parentTag: selectedStory.parentTag,
       onClick: async () => {
-        if (currentTimeout) {
-          clearTimeout(currentTimeout);
-          currentTimeout = null;
+        if (timeout) {
+          clearTimeout(timeout);
+          timeout = null;
         }
         if (!toasterInstance) {
           createToaster();
         } else {
+          if (activeToasters.closeIfExists(`${variant.h}-${variant.v}`)) {
+            return;
+          }
           const toasterConfig = {
             closeButton: true,
             anchorOrigin: { horizontal: variant.h, vertical: variant.v },
@@ -2145,19 +2165,23 @@ function storyToaster(selectedStory) {
                              <span style="font-size: 1rem; font-weight: 600;">Title:</span>
                              <span style="font-size: 0.85rem; font-weight: 400;">toaster: ${variant.h}-${variant.v}</span>
                            </div>`,
-            autoHideDuration: 1e3,
+            autoHideDuration: variant.autoHideDuration,
             type: variant.type,
             size: variant.size,
             animation: { name: variant.animation },
             persist: true,
             onComplete: (context) => {
-              currentTimeout = setTimeout(() => {
-                context.Toaster.close();
-                currentTimeout = null;
-              }, context.Toaster.autoHideDuration());
+            },
+            onClosed: (context) => {
+              console.log("Toaster Closed");
+              delete activeToasters[`${variant.h}-${variant.v}`];
+            },
+            onClose: (context) => {
+              console.log("Button Clicked");
+              context.Toaster.close();
             }
           };
-          await toasterInstance.restart(toasterConfig);
+          activeToasters[`${variant.h}-${variant.v}`] = await toasterInstance.restart(toasterConfig);
         }
       }
     };